Creates get users hook
